Conflicts:
	conf/reporting.conf
	data/html/sections/behavior.html
	data/html/sections/signatures.html
	data/html/sections/volatility.html
	docs/book/src/installation/host/configuration.rst
	modules/reporting/reporthtml.py
	requirements.txt
	utils/web.py

+++ lib/cuckoo/core/startup.py
@@ -1,4 +1,5 @@
-# Copyright (C) 2010-2015 Cuckoo Foundation.
+# Copyright (C) 2010-2013 Claudio Guarnieri.
+# Copyright (C) 2014-2016 Cuckoo Foundation.
 # This file is part of Cuckoo Sandbox - http://www.cuckoosandbox.org
 # See the file 'docs/LICENSE' for copying permission.
 
@@ -11,21 +12,24 @@ import urllib
 import urllib2
 import logging
 import logging.handlers
-import pwd
-
-import modules.auxiliary
-import modules.processing
-import modules.signatures
-import modules.reporting
+import socket
 
 from lib.cuckoo.common.colors import red, green, yellow, cyan
 from lib.cuckoo.common.config import Config
 from lib.cuckoo.common.constants import CUCKOO_ROOT, CUCKOO_VERSION
-from lib.cuckoo.common.exceptions import CuckooStartupError
+from lib.cuckoo.common.exceptions import CuckooStartupError, CuckooDatabaseError
 from lib.cuckoo.common.exceptions import CuckooOperationalError
 from lib.cuckoo.common.utils import create_folders
-from lib.cuckoo.core.database import Database, TASK_RUNNING, TASK_FAILED_ANALYSIS
+from lib.cuckoo.core.database import Database, TASK_RUNNING
+from lib.cuckoo.core.database import TASK_FAILED_ANALYSIS, TASK_PENDING
 from lib.cuckoo.core.plugins import import_plugin, import_package, list_plugins
+from lib.cuckoo.core.rooter import rooter, vpns
+
+try:
+    import pwd
+    HAVE_PWD = True
+except ImportError:
+    HAVE_PWD = False
 
 log = logging.getLogger()
 
@@ -56,11 +60,12 @@ def check_configs():
     """Checks if config files exist.
     @raise CuckooStartupError: if config files do not exist.
     """
-    configs = [os.path.join(CUCKOO_ROOT, "conf", "cuckoo.conf"),
-               os.path.join(CUCKOO_ROOT, "conf", "reporting.conf"),
-               os.path.join(CUCKOO_ROOT, "conf", "auxiliary.conf")]
+    configs = ("auxiliary.conf", "avd.conf", "cuckoo.conf", "esx.conf",
+               "kvm.conf", "memory.conf", "physical.conf", "processing.conf",
+               "qemu.conf", "reporting.conf", "virtualbox.conf", "vmware.conf",
+               "vpn.conf", "vsphere.conf", "xenserver.conf")
 
-    for config in configs:
+    for config in [os.path.join(CUCKOO_ROOT, "conf", f) for f in configs]:
         if not os.path.exists(config):
             raise CuckooStartupError("Config file does not exist at "
                                      "path: {0}".format(config))
@@ -73,7 +78,8 @@ def create_structure():
         "log",
         "storage",
         os.path.join("storage", "analyses"),
-        os.path.join("storage", "binaries")
+        os.path.join("storage", "binaries"),
+        os.path.join("storage", "baseline"),
     ]
 
     try:
@@ -101,16 +107,17 @@ def check_version():
         return
 
     try:
-        response_data = json.loads(response.read())
+        r = json.loads(response.read())
     except ValueError:
         print(red(" Failed! ") + "Invalid response.\n")
         return
 
-    if not response_data["error"]:
-        if response_data["response"] == "NEW_VERSION":
-            msg = "Cuckoo Sandbox version {0} is available " \
-                  "now.\n".format(response_data["current"])
+    if not r["error"]:
+        if r["response"] == "NEW_VERSION" and r["current"] != "1.2":
+            msg = "Cuckoo Sandbox version %s is available now." % r["current"]
             print(red(" Outdated! ") + msg)
+        elif r["current"] == "1.2":
+            print(yellow(" Okay! ") + "You are running a development version.")
         else:
             print(green(" Good! ") + "You have the latest version "
                                      "available.\n")
@@ -177,10 +184,8 @@ def init_tasks():
     db = Database()
     cfg = Config()
 
-    log.debug("Checking for locked tasks...")
-    tasks = db.list_tasks(status=TASK_RUNNING)
-
-    for task in tasks:
+    log.debug("Checking for locked tasks..")
+    for task in db.list_tasks(status=TASK_RUNNING):
         if cfg.cuckoo.reschedule:
             db.reschedule(task.id)
             log.info("Rescheduled task with ID {0} and "
@@ -189,21 +194,51 @@ def init_tasks():
             db.set_status(task.id, TASK_FAILED_ANALYSIS)
             log.info("Updated running task ID {0} status to failed_analysis".format(task.id))
 
-def init_modules():
+    log.debug("Checking for pending service tasks..")
+    for task in db.list_tasks(status=TASK_PENDING, category="service"):
+        db.set_status(task.id, TASK_FAILED_ANALYSIS)
+
+def delete_file(*rel_path):
+    filepath = os.path.join(CUCKOO_ROOT, *rel_path)
+    if not os.path.exists(filepath):
+        return
+
+    try:
+        os.unlink(filepath)
+    except Exception as e:
+        log.warning(
+            "Unable to remove old %s leftover file from before you updated "
+            "your Cuckoo setup to the latest version: %s.",
+            os.path.join(*rel_path), e
+        )
+
+def init_modules(machinery=True):
     """Initializes plugins."""
     log.debug("Importing modules...")
 
     # Import all auxiliary modules.
+    import modules.auxiliary
     import_package(modules.auxiliary)
+
     # Import all processing modules.
+    import modules.processing
     import_package(modules.processing)
+
     # Import all signatures.
+    import modules.signatures
     import_package(modules.signatures)
+
+    delete_file("modules", "reporting", "maec40.pyc")
+    delete_file("modules", "reporting", "maec41.pyc")
+    delete_file("modules", "reporting", "mmdef.pyc")
+
     # Import all reporting modules.
+    import modules.reporting
     import_package(modules.reporting)
 
     # Import machine manager.
-    import_plugin("modules.machinery." + Config().cuckoo.machinery)
+    if machinery:
+        import_plugin("modules.machinery." + Config().cuckoo.machinery)
 
     for category, entries in list_plugins().items():
         log.debug("Imported \"%s\" modules:", category)
@@ -266,11 +301,160 @@ def init_yara():
         else:
             log.debug("\t |-- %s", entry)
 
+def init_binaries():
+    """Inform the user about the need to periodically look for new analyzer
+    binaries. These include the Windows monitor etc."""
+    monitor = os.path.join(CUCKOO_ROOT, "data", "monitor", "latest")
+
+    # Checks whether the "latest" symlink is available as well as whether
+    # it points to an existing directory.
+    if not os.path.exists(monitor):
+        raise CuckooStartupError(
+            "The binaries used for Windows analysis are updated regularly, "
+            "independently from the release line. It appears that you're "
+            "not up-to-date. This can happen when you've just installed "
+            "Cuckoo or when you've updated your Cuckoo version by pulling "
+            "the latest changes from our Git repository. In order to get "
+            "up-to-date, please run the following "
+            "command: `./utils/community.py -wafb monitor` or "
+            "`./utils/community.py -wafb 2.0` if you'd also like to download "
+            "over 300 Cuckoo signatures."
+        )
+
+def init_rooter():
+    """If required, check whether the rooter is running and whether we can
+    connect to it."""
+    cuckoo = Config()
+
+    # The default configuration doesn't require the rooter to be ran.
+    if not Config("vpn").vpn.enabled and cuckoo.routing.route == "none":
+        return
+
+    cuckoo = Config()
+    s = socket.socket(socket.AF_UNIX, socket.SOCK_DGRAM)
+
+    try:
+        s.connect(cuckoo.cuckoo.rooter)
+    except socket.error as e:
+        if e.strerror == "No such file or directory":
+            raise CuckooStartupError(
+                "The rooter is required but it is either not running or it "
+                "has been configured to a different Unix socket path. "
+                "(In order to disable the use of rooter, please set route "
+                "and internet to none in cuckoo.conf and enabled to no in "
+                "vpn.conf)."
+            )
+
+        if e.strerror == "Connection refused":
+            raise CuckooStartupError(
+                "The rooter is required but we can't connect to it as the "
+                "rooter is not actually running. "
+                "(In order to disable the use of rooter, please set route "
+                "and internet to none in cuckoo.conf and enabled to no in "
+                "vpn.conf)."
+            )
+
+        if e.strerror == "Permission denied":
+            raise CuckooStartupError(
+                "The rooter is required but we can't connect to it due to "
+                "incorrect permissions. Did you assign it the correct group? "
+                "(In order to disable the use of rooter, please set route "
+                "and internet to none in cuckoo.conf and enabled to no in "
+                "vpn.conf)."
+            )
+
+        raise CuckooStartupError("Unknown rooter error: %s" % e)
+
+    # Do not forward any packets unless we have explicitly stated so.
+    rooter("forward_drop")
+
+def init_routing():
+    """Initialize and check whether the routing information is correct."""
+    cuckoo = Config()
+    vpn = Config("vpn")
+
+    # Check whether all VPNs exist if configured and make their configuration
+    # available through the vpns variable. Also enable NAT on each interface.
+    if vpn.vpn.enabled:
+        for name in vpn.vpn.vpns.split(","):
+            name = name.strip()
+            if not name:
+                continue
+
+            if not hasattr(vpn, name):
+                raise CuckooStartupError(
+                    "Could not find VPN configuration for %s" % name
+                )
+
+            entry = vpn.get(name)
+
+            if not rooter("nic_available", entry.interface):
+                raise CuckooStartupError(
+                    "The network interface that has been configured for "
+                    "VPN %s is not available." % entry.name
+                )
+
+            if not rooter("rt_available", entry.rt_table):
+                raise CuckooStartupError(
+                    "The routing table that has been configured for "
+                    "VPN %s is not available." % entry.name
+                )
+
+            vpns[entry.name] = entry
+
+            # Disable & enable NAT on this network interface. Disable it just
+            # in case we still had the same rule from a previous run.
+            rooter("disable_nat", entry.interface)
+            rooter("enable_nat", entry.interface)
+
+            # Populate routing table with entries from main routing table.
+            if cuckoo.routing.auto_rt:
+                rooter("flush_rttable", entry.rt_table)
+                rooter("init_rttable", entry.rt_table, entry.interface)
+
+    # Check whether the default VPN exists if specified.
+    if cuckoo.routing.route not in ("none", "internet"):
+        if not vpn.vpn.enabled:
+            raise CuckooStartupError(
+                "A VPN has been configured as default routing interface for "
+                "VMs, but VPNs have not been enabled in vpn.conf"
+            )
+
+        if cuckoo.routing.route not in vpns:
+            raise CuckooStartupError(
+                "The VPN defined as default routing target has not been "
+                "configured in vpn.conf."
+            )
+
+    # Check whether the dirty line exists if it has been defined.
+    if cuckoo.routing.internet != "none":
+        if not rooter("nic_available", cuckoo.routing.internet):
+            raise CuckooStartupError(
+                "The network interface that has been configured as dirty "
+                "line is not available."
+            )
+
+        if not rooter("rt_available", cuckoo.routing.rt_table):
+            raise CuckooStartupError(
+                "The routing table that has been configured for dirty "
+                "line interface is not available."
+            )
+
+        # Disable & enable NAT on this network interface. Disable it just
+        # in case we still had the same rule from a previous run.
+        rooter("disable_nat", cuckoo.routing.internet)
+        rooter("enable_nat", cuckoo.routing.internet)
+
+        # Populate routing table with entries from main routing table.
+        if cuckoo.routing.auto_rt:
+            rooter("flush_rttable", cuckoo.routing.rt_table)
+            rooter("init_rttable", cuckoo.routing.rt_table,
+                   cuckoo.routing.internet)
 
 def cuckoo_clean():
     """Clean up cuckoo setup.
-    It deletes logs, all stored data from file system and configured databases (SQL
-    and MongoDB.
+    It deletes logs, all stored data from file system and configured
+    databases (SQL and MongoDB).
     """
     # Init logging.
     # This need to init a console logger handler, because the standard
@@ -279,10 +463,17 @@ def cuckoo_clean():
     init_console_logging()
 
     # Initialize the database connection.
-    db = Database()
-
-    # Drop all tables.
-    db.drop()
+    try:
+        db = Database(schema_check=False)
+    except CuckooDatabaseError as e:
+        # If something is screwed due to incorrect database migrations or bad
+        # database SqlAlchemy would be unable to connect and operate.
+        log.warning("Error connecting to database: it is suggested to check "
+                    "the connectivity, apply all migrations if needed or purge "
+                    "it manually. Error description: %s", e)
+    else:
+        # Drop all tables.
+        db.drop()
 
     # Check if MongoDB reporting is enabled and drop that if it is.
     cfg = Config("reporting")
@@ -319,7 +510,12 @@ def cuckoo_clean():
             if not fname.endswith(".pyc"):
                 continue
 
-            path = os.path.join(CUCKOO_ROOT, dirpath, fname)
+            # We don't want to delete the Android's Agent .pyc files (as we
+            # don't ship the original .py files and thus they're critical).
+            if "agent/android/python_agent" in dirpath.replace("\\", "/"):
+                continue
+
+            path = os.path.join(dirpath, fname)
 
             try:
                 os.unlink(path)
@@ -330,6 +526,10 @@ def drop_privileges(username):
     """Drops privileges to selected user.
     @param username: drop privileges to this username
     """
+    if not HAVE_PWD:
+        sys.exit("Unable to import pwd required for dropping "
+                 "privileges (`pip install pwd`)")
+
     try:
         user = pwd.getpwnam(username)
         os.setgroups((user.pw_gid,))
